BCDA-9911: Add context checks for worker jobs (#1339)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9911

## 🛠 Changes

Add context checks for worker jobs. I think my biggest concern here is
adding max timeouts to all our worker processes:
- default: 10m (applies to Prepare job and Cleanup job)
- specified for Process job: 30m

## ℹ️ Context

Currently worker processes dont respect parent context, ie
cancellations, timeouts, etc. This can result in jobs getting "stuck"
and needing to be recovered after an hour by a separate process.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Local linting and testing.

---------
