+++ bcdaworker/queueing/river.go
@@ -21,7 +21,6 @@ import (
 	"time"
 
 	"github.com/CMSgov/bcda-app/bcda/database"
-	"github.com/CMSgov/bcda-app/bcda/utils"
 	"github.com/CMSgov/bcda-app/bcdaworker/queueing/worker_types"
 	"github.com/CMSgov/bcda-app/bcdaworker/repository/postgres"
 	"github.com/CMSgov/bcda-app/bcdaworker/worker"
@@ -71,10 +70,9 @@ func StartRiver(db *sql.DB, numWorkers int) *queue {
 		Queues: map[string]river.QueueConfig{
 			river.QueueDefault: {MaxWorkers: numWorkers},
 		},
-		// TODO: whats an appropriate timeout?
-		JobTimeout:   -1, // default for river is 1m, using -1 for no timeout
+		JobTimeout:   10 * time.Minute,
 		Logger:       logger,
-		MaxAttempts:  10, // This is a few hours worth of retries
+		MaxAttempts:  8, // This is roughly an hour of retries
 		Workers:      workers,
 		PeriodicJobs: periodicJobs,
 	})
@@ -103,9 +101,4 @@ func (q queue) StopRiver() {
 	if err := q.client.Stop(q.ctx); err != nil {
 		panic(err)
 	}
-}
-
-func getCutOffTime() time.Time {
-	cutoff := time.Now().Add(-time.Hour * time.Duration(utils.GetEnvInt("ARCHIVE_THRESHOLD_HR", 24)))
-	return cutoff
 }
